Move PowerGoggles classes to gregtech.common.powergoggles
