+++ src/main/java/gtPlusPlus/core/proxy/ClientProxy.java
@@ -21,9 +21,9 @@ import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.event.FMLServerStartingEvent;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
-import gregtech.common.handlers.PowerGogglesHudHandler;
-import gregtech.common.handlers.PowerGogglesKeybindHandler;
-import gregtech.common.items.gui.PowerGogglesGuiOverlay;
+import gregtech.common.powergoggles.handlers.PowerGogglesHudHandler;
+import gregtech.common.powergoggles.handlers.PowerGogglesKeybindHandler;
+import gregtech.common.powergoggles.gui.PowerGogglesGuiOverlay;
 import gtPlusPlus.GTplusplus;
 import gtPlusPlus.core.client.renderer.CustomItemBlockRenderer;
 import gtPlusPlus.core.client.renderer.CustomOreBlockRenderer;
